Thread 64-bit folder sizes through start/send transfers
startFolderUpload size becomes UInt64 (protocol + impl); request adds
xferSize64 for >4 GiB folders alongside the clamped legacy field.
startFolderDownload reads xferSize64 with a UInt32 fallback. Folder
send/decode get largeFile from the session's largeFilesEnabled.

+++ Sources/HeidrunCore/HotlineClient.swift
@@ -393,11 +393,12 @@ public protocol HotlineClient: Sendable {
         resume: Bool
     ) async throws -> TransferHandle
 
-    /// Begin uploading a folder.
+    /// Begin uploading a folder. `size` is the folder's total payload in
+    /// bytes; it may exceed 4 GiB when the session negotiated large files.
     func startFolderUpload(
         at path: RemotePath,
         name: String,
-        size: UInt32,
+        size: UInt64,
         itemCount: UInt16,
         resume: Bool
     ) async throws -> TransferHandle
